transition: Allow specifying the diffing algorithm
For example if you need to take into account live TV behind a
translucent background, or noise / encoding artifacts that continue to
change after the transition is completed. This doesn't actually add an
implementation for such cases, but it makes it possible to do without
needing a new stb-tester release.