Add stbt.MultiPress: Enter text using a numeric keypad
In some apps, the search page allows entering text by pressing the keys
on the remote control's numeric keypad: press the number "2" once for
"A", twice for "B", etc.::

        1.,     ABC2    DEF3
        GHI4    JKL5    MNO6
        PQRS7   TUV8    WXYZ9
              [space]0

This is much easier (and faster) than moving a selection around an
on-screen keyboard.

The keys for A-Z are specified in [ITU E.161]. The standard doesn't
specify where "space" goes, but every implementation I have seen has it
on KEY_0. Punctuation varies across implementations; I have arbitrarily
chosen the behaviour of a Sky HD+ box that I had available for testing,
but you can override this by passing `key_mapping={"KEY_1": "@1.,-_}` to
the `stbt.MultiPress` constructor.

[ITU E.161]: https://www.itu.int/rec/T-REC-E.161