stbt.Keyboard: Change internal graph representation to support "modes"
The internal format of the graph (and the graph itself) are now private
implementation details. This is a breaking change.

Nodes in the graph are now (name, text, region, mode). Some of those may
be `None`. To look up a node in the graph you need to provide at least
one of those (sometimes you need to provide more than one if you want to
find an unambiguous node). This will allow simpler PageObject code:
the PageObject won't need to map from selection region to text, if the
region is already stored in the Keyboard model.

Now that the format of the directed graph (and the graph itself) are
private implementation details, I have changed the user-facing
terminology to be "key" for the on-screen keys on the keyboard (nodes),
and "transition" for the edges. This is slightly confusing because
typically we use "key" to mean the name of a key on the remote control;
here I've used "keypress" (which elsewhere we use to mean the action of
pressing a key, e.g. the return value from `stbt.press`). In the code I
still use "node" and "edge" when I'm in a low-level private function.

I've had to keep the `graph` parameter to `stbt.Keyboard`, because
removing it would mean that people who were passing it as a positional
argument will now be setting the `mask`, and get a very confusing error
message. This way the error will be much clearer, and we can remove the
parameter altogether in v33.